+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/auth/compare/2.0.2...v2.x)
 
+### Added
+
+- `Firewall`
+	- callbacks are marked as `@param-later-invoked-callable`
+
 ## [2.0.2](https://github.com/orisai/auth/compare/2.0.1...2.0.2) - 2024-03-08
 
 ### Fixed