+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Firewall`
 	- callbacks are marked as `@param-later-invoked-callable`
 
+### Changed
+
+- Allow PHP 8.3
+
 ## [2.0.2](https://github.com/orisai/auth/compare/2.0.1...2.0.2) - 2024-03-08
 
 ### Fixed